+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # PLUTO Change Log
 
+## 1.0.7
+### New Features
+
+#### Validation Summary
+Added a new page off of the validations list from the "Summary" button, which shows a summary of the validation.
+
+#### Validation Export/Import
+Added the ability for a PLUTO instance to export validations to another PLUTO instance.
+
+### Bug Fixes and Minor Improvements
+ * added a docker reference stack (with logging, performance metrics, ElasticSearch, and Kibana)
+ * Cleaned up logging from the web service to allow easier ingestion by Filebeat
+ * Added run summary details to web service output log
+ * Fix bug on disabled create validation continue button
+ * Cleaned up layout of config fields in UI
+ * Hide the Additional Information if there are none configured
+ * Added limit on number of concurrent running validations in a service
+ * Added environment variable replacement support to all strings inside validatorConfig.json
+
 ## 0.9.6
 ### New Features